+++ pkg/ingester/ingester_test.go
@@ -715,7 +715,7 @@ func Test_DedupeIngester(t *testing.T) {
 		it := iter.NewMergeSampleIterator(ctx, iterators)
 		var expectedLabels []string
 		for _, s := range streams {
-			expectedLabels = append(expectedLabels, labels.NewBuilder(s).Del("foo").Labels().String())
+			expectedLabels = append(expectedLabels, labels.NewBuilder(s).Del("foo").Labels(nil).String())
 		}
 		sort.Strings(expectedLabels)
 		for i := int64(0); i < requests; i++ {