+++ pkg/ingester/instance_test.go
@@ -478,7 +478,7 @@ func makeRandomLabels() labels.Labels {
 	for _, ln := range labelNames {
 		ls.Set(ln, fmt.Sprintf("%d", rand.Int31()))
 	}
-	return ls.Labels()
+	return ls.Labels(nil)
 }
 
 func Benchmark_PushInstance(b *testing.B) {